java: Use SERVICE_NOT_FOUND for NATS
As of nats-server 2.2.0, a PUB sent to a subject with no responders
results in an HMSG status message with NO_RESPONDERS_CODE and no data
being sent to the reply subject.  The NATS server does not send any
other metadata to identify the request, so change the inbox subscription
to a wildcard, and encode the op id in the reply subject.