dart: Fix use_null_for_unset,nullsafe
For non-required fields, use_null_for_unset generates isSet() checks
(except for internal non-exported structs), which works but means the
Dart analyzer can't see the null checks.  Instead, prefer null checks
when the isSet() method is just a null check anyway, which is always the
case for struct/list/set/map.  In practice, isSet()-vs-null only matters
for struct (oproto.writeX) and list/set/map (list.length and for-loop)
since all other types are passed to null-tolerant oprot.writeX methods.

For required struct/list/set/map, the validate() method called at the
top of the generate write() method already ensures the fields are
non-null, so use_null_for_unset skips the redundant null check, but the
compiler can't see that, so we need to force non-null.  Pre-existing,
the same null check is redundant for non-use_null_for_unset too, so the
code is adjusted to ensure a non-null assertion would be added if that
change is made in the future.